requirement.txt
numpy   
pandas
sqlalchemy
pyodbc
python-dotenv
scikit-learn
logging